chore: fix ui routing
